Simplify kvariants module structure
